+++ examples/dec_frame.rs
@@ -2,6 +2,7 @@
 use std::env;
 use std::error::Error;
 use std::fs::File;
+use std::io::BufReader;
 
 fn main() -> Result<(), Box<dyn Error>> {
     let path = env::args().nth(1).expect("usage: dec_frame [filename]");
@@ -9,7 +10,7 @@ fn main() -> Result<(), Box<dyn Error>> {
 }
 
 fn decode(path: &str) -> Result<(), Box<dyn Error>> {
-    let f = File::open(path)?;
+    let f = BufReader::new(File::open(path)?);
     let mut frame_dec = gift::Decoder::new(f).into_frames();
     let preamble = frame_dec.preamble()?;
     println!("preamble: {:?}", preamble);