+++ examples/dec_raster.rs
@@ -2,6 +2,7 @@
 use std::env;
 use std::error::Error;
 use std::fs::File;
+use std::io::BufReader;
 
 fn main() -> Result<(), Box<dyn Error>> {
     let path = env::args().nth(1).expect("usage: dec_raster [filename]");
@@ -9,7 +10,7 @@ fn main() -> Result<(), Box<dyn Error>> {
 }
 
 fn decode(path: &str) -> Result<(), Box<dyn Error>> {
-    let gif = File::open(path)?;
+    let gif = BufReader::new(File::open(path)?);
     for step in gift::Decoder::new(gif) {
         let step = step?;
         let raster = step.raster();